#WE-10 Проверка валидности данных Prof.xaml

+++ ToTaxi/Prof.xaml.cs
@@ -16,6 +16,9 @@ using System.Net;
 using Image = System.Drawing.Image;
 using System.Windows.Media.Imaging;
 using Microsoft.Win32;
+using System.Windows.Media;
+using System.Text.RegularExpressions;
+using System.Globalization;
 
 namespace ToTaxi
 {
@@ -43,32 +46,38 @@ namespace ToTaxi
             {
                 var s = y.Users.Where(p => p.Id == Global._ID);
                 foreach (var f in s)
-                {
-                    using (MemoryStream ms = new MemoryStream(f.Photo))
+                    if(f.Photo == null)
+                    { GoBack(); }
+                    else
                     {
-                        using (Image img = Image.FromStream(ms))
+
                         {
-                            try
+                            using (MemoryStream ms = new MemoryStream(f.Photo))
                             {
-                                img.Save($@"Img/{f.Id}.Png");
-                                var path = $@"Img/{f.Id}.Png";
-                                Bitmap bmp = new Bitmap(path);
-                                MemoryStream ms1 = new MemoryStream();
-                                bmp.Save(ms1, ImageFormat.Png);
-                                ms1.Position = 0;
-                                BitmapImage bi = new BitmapImage();
-                                bi.BeginInit();
-                                bi.StreamSource = ms1;
-                                bi.EndInit();
-                                w1.Source = bi;
-                            }
-                            catch
-                            {
-                                File.Delete($@"Img/{f.Id}.Png");
-                                FOIDG();
+                                using (Image img = Image.FromStream(ms))
+                                {
+                                    try
+                                    {
+                                        img.Save($@"Img/{f.Id}.Png");
+                                        var path = $@"Img/{f.Id}.Png";
+                                        Bitmap bmp = new Bitmap(path);
+                                        MemoryStream ms1 = new MemoryStream();
+                                        bmp.Save(ms1, ImageFormat.Png);
+                                        ms1.Position = 0;
+                                        BitmapImage bi = new BitmapImage();
+                                        bi.BeginInit();
+                                        bi.StreamSource = ms1;
+                                        bi.EndInit();
+                                        w1.Source = bi;
+                                    }
+                                    catch
+                                    {
+                                        File.Delete($@"Img/{f.Id}.Png");
+                                        FOIDG();
+                                    }
+                                }
                             }
                         }
-                    }
 
                     GoBack();
 
@@ -157,38 +166,96 @@ namespace ToTaxi
 
         public void GoSave()
         {
-            fam.IsReadOnly = true;
-            im.IsReadOnly = true;
-            ot.IsReadOnly = true;
-            eml.IsReadOnly = true;
-            tel.IsReadOnly = true;
-            dat.IsEnabled = false;
-            pas.IsEnabled = false;
-            poll.IsReadOnly = true;
-            using (TaxiInDronContext v = new TaxiInDronContext())
+            StringBuilder B = new StringBuilder();
+            if (pas.Password.Length < 6 )
             {
-                var g = v.Users.Where(p => p.Id == Global._ID).SingleOrDefault();
-                g.Name = im.Text;
-                g.Fam = fam.Text;
-                g.Otc = ot.Text;
-                g.PasswordInVx = pas.Password;
-                g.Tel = tel.Text;
-                g.DateBird0 = dat.SelectedDate;
-                g.Email = eml.Text;
-                g.Photo = fot;
-                if (m1.IsChecked == true)
+                B.Append("Пароль слишком короткий\n");
+            }
+            if(pas.Password.Any(char.IsUpper) == false)
+            {
+                B.Append("Пароль должен содержать хотябы одну букву верхнего регистра\n");
+            }
+            if(pas.Password.Any(Char.IsLower) == false)
+            {
+                B.Append("Пароль должен содержать хотябы одну букву нижнего регистра\n");
+            }
+            if(IsValidEmail(eml.Text) == false )
+            {
+               B.Append(" Введеите корректнкю почту\n");
+            }
+            if (B.Length == 0)
+            {
+                fam.IsReadOnly = true;
+                im.IsReadOnly = true;
+                ot.IsReadOnly = true;
+                eml.IsReadOnly = true;
+                tel.IsReadOnly = true;
+                dat.IsEnabled = false;
+                pas.IsEnabled = false;
+                poll.IsReadOnly = true;
+                using (TaxiInDronContext v = new TaxiInDronContext())
                 {
-                    g.Pol = 10;
+                    var g = v.Users.Where(p => p.Id == Global._ID).SingleOrDefault();
+                    g.Name = im.Text;
+                    g.Fam = fam.Text;
+                    g.Otc = ot.Text;
+                    g.PasswordInVx = pas.Password;
+                    g.Tel = tel.Text;
+                    g.DateBird0 = dat.SelectedDate;
+                    g.Email = eml.Text;
+                    g.Photo = fot;
+                    if (m1.IsChecked == true)
+                    {
+                        g.Pol = 10;
+                    }
+                    if (m2.IsChecked == true)
+                    {
+                        g.Pol = 11;
+                    }
+                    v.SaveChanges();
+                    FOIDG();
                 }
-                if (m2.IsChecked == true)
-                {
-                    g.Pol = 11;
-                }
-                v.SaveChanges();
-                FOIDG();
+            }
+            else
+            {
+                MessageBox.Show(B.ToString());
             }
 
 
+        }
+        public static bool IsValidEmail(string email)
+        {
+            if (string.IsNullOrWhiteSpace(email))
+                return false;
+            try
+            {
+                email = Regex.Replace(email, @"(@)(.+)$", DomainMapper,
+                                      RegexOptions.None, TimeSpan.FromMilliseconds(200));
+                string DomainMapper(Match match)
+                {
+                    var idn = new IdnMapping();
+                    string domainName = idn.GetAscii(match.Groups[2].Value);
+                    return match.Groups[1].Value + domainName;
+                }
+            }
+            catch (RegexMatchTimeoutException e)
+            {
+                return false;
+            }
+            catch (ArgumentException e)
+            {
+                return false;
+            }
+            try
+            {
+                return Regex.IsMatch(email,
+                    @"^[^@\s]+@[^@\s]+\.[^@\s]+$",
+                    RegexOptions.IgnoreCase, TimeSpan.FromMilliseconds(250));
+            }
+            catch (RegexMatchTimeoutException)
+            {
+                return false;
+            }
         }
 
         private void f1_Copy_Click(object sender, RoutedEventArgs e)
@@ -215,7 +282,15 @@ namespace ToTaxi
             m2.IsChecked = false;
         }
 
-       
+        private void ls_Click(object sender, RoutedEventArgs e)
+        {
+            Fram.MainFF.Navigate(new Prof());
+        }
+
+        private void Rol_Click(object sender, RoutedEventArgs e)
+        {
+
+        }
     }
 }
